Update README.md
This is what I have and I am trying to learn how to integrate the commented out to the main code to be functional. I am trying to add the option record high score data on the GUI. I want to continue to build on this to increasing speed to the object drop every ten lines.